This is a snapshot commit which is the merge of the official 'v1.3.3'
release into the STMicroelectronics' ST40 tree. Basically, it is a
re-baselining activity from 'v1.3.2' to 'v1.3.3'.

This commit is largely just the automatic "merge", together with the
manual merge of the remaining conflicts. At least for ST40 boards,
this snapshot builds cleanly, and superficially appears to run on the
one board (HDK-7208) tested.

There is one main functional delinquency from this commit:

1) common/cmd_bootm.c and lib_st40/bootm.c

	Support for the "bootm" command had *not* been updated, to take
	into account the new "image" infrastructure. This clearly needs
	to be added, but probably best done as a separate commit.

This delinquency should be fully addressed in the next few commits.

Conflicts:
	MAKEALL
	Makefile
	README
	common/cmd_bootm.c
	common/cmd_ide.c
	common/cmd_nand.c
	drivers/mtd/cfi_flash.c
	drivers/net/Makefile
	drivers/usb/usb_ohci.c
	include/common.h
	include/image.h
	include/nand.h
	lib_generic/crc32.c
	lib_generic/sha1.c
	tools/mkimage.c

+++ include/u-boot/zlib.h
@@ -312,7 +312,7 @@ extern "C" {
 typedef unsigned char  Byte;  /* 8 bits */
 #endif
 typedef unsigned int   uInt;  /* 16 bits or more */
-typedef unsigned long  uLong; /* 32 bits or more */
+typedef unsigned int   uLong; /* 32 bits or more */
 
 #ifdef SMALL_MEDIUM
    /* Borland C/C++ and some old MSC versions ignore FAR inside typedef */
